move quiz script and add jinja
