+++ CONTRIBUTING.md
@@ -284,13 +284,13 @@ If you would like to create a new figure, let us know and we can provide access 
 
 ### Create custom quizzes and flashcards
 
-To build quizzes or flashcards, use the helper functions in `jupyter-book/src/lib.py`.
+To build quizzes or flashcards, use the helper functions in `scripts/quiz.py`.
 You can create multiple-choice questions or simple flip cards.
 
 1. Start your notebook code cell with
 
 ```python
-%run ../src/lib.py
+%run ../../scripts/quiz.py
 ```
 
 2. Then, add as many questions as you like. For example:
@@ -316,7 +316,7 @@ multiple_choice_question(
 > Add the cell tag `remove-input` to the code cell to remove the code, when building the book.
 
 You can also adjust the font size, text color and much more.
-For detailed information, check out the method descriptions in `jupyter-book/src/lib.py`.
+For detailed information, check out the method descriptions in `scripts/quiz.py`.
 
 ### Pre commit
