+++ environment.yml
@@ -7,6 +7,7 @@ dependencies:
   - conda-forge::jupyter-book==2.1.6
   - conda-forge::jupytext==1.19.3
   - conda-forge::beautifulsoup4==4.15.0
+  - conda-forge::jinja2
   - pip
   - pip:
       - lamindb
